+++ BaseDeProjetos/Models/ProjetoDTO.cs
@@ -69,6 +69,6 @@ namespace BaseDeProjetos.Models
 
         public virtual float? SatisfacaoClienteFinal { get; set; }
 
-        public virtual float CustoHH { get; set; }
+        public virtual decimal CustoHH { get; set; }
     }
 }
